% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatenate_vectors_py.R
\name{concatenate_vectors_py}
\alias{concatenate_vectors_py}
\title{Add or Concatenate Objects via Python}
\usage{
concatenate_vectors_py(x, y)
}
\arguments{
\item{x}{Numeric or vector}

\item{y}{Numeric or vector to be added to/combined with}
}
\value{
The result of the Python add_numbers function depending on the input types.
}
\description{
A wrapper function that passes two R objects to a Python script.
Depending on the input types, Python will either sum them (if numeric)
or concatenate them (if vectors/strings).
}
